bugfix: minor changes to mallocChar()
casting to (unsigned long) may be incorrect between different architectures.

+++ toolsMemory.c
@@ -66,7 +66,7 @@ char*				mallocChar(uintptr_t numChars){
 	*/
 	
 	char*	temp = NULL;	//temporary pointer
-	temp = malloc((unsigned long)numChars*sizeof(char));
+	temp = malloc((size_t)numChars*sizeof(char));
 	if (temp == NULL){
 		fatal("Memory allocation error.\n");
 	}
